fix(mpi): preserve dtype in localize and cover vector/empty/float32 cases
- localize / _dof_coeffs no longer force float64: cast to the global space's
  dtype so float32 Bspline spaces work (Bspline requires an exact dtype match).
- tests: vector Bspline through create_distributed_function (exercises the
  _dof_coeffs flatten path), empty-rank DistributedFunction, and a float32 case.

+++ src/pantr/mpi/_distributed_space.py
@@ -226,7 +226,7 @@ class DistributedSpace:
         local = self._local
         if local is None:
             return None
-        cp = np.asarray(control_points, dtype=np.float64)
+        cp = np.asarray(control_points, dtype=self._global_space.dtype)
         n_global = self._global_space.num_total_basis
         if cp.shape[0] != n_global:
             raise ValueError(
